Move CppSignature and CppSignatureGroup into cpp, add __all__
It was pretty weird that CppSignature lived in types, so this moves
it to the right place.  Once we do this, cpp is better encapsulated,
so we can in fact turn off most exports except for CppSignature and
CppSignatureGroup


[ghstack-poisoned]
