+++ tools/autograd/gen_python_functions.py
@@ -43,6 +43,7 @@ from tools.codegen.api.python import *
 from tools.codegen.gen import cpp_string, parse_native_yaml, with_native_function, FileManager
 from tools.codegen.model import *
 from tools.codegen.utils import *
+from tools.codegen.api.cpp import CppSignatureGroup
 
 from typing import Dict, Optional, List, Tuple, Set, Sequence, Callable
