+++ tools/autograd/load_derivatives.py
@@ -10,6 +10,7 @@ import yaml
 from tools.codegen.api.autograd import *
 from tools.codegen.api.types import *
 import tools.codegen.api.cpp as cpp
+from tools.codegen.api.cpp import CppSignatureGroup
 from tools.codegen.gen import parse_native_yaml, with_native_function
 from tools.codegen.model import *
 from tools.codegen.utils import *
