+++ tools/codegen/api/python.py
@@ -3,6 +3,7 @@ from typing import Optional, Union, Sequence, Set, List, Dict, Tuple
 
 from tools.codegen.api.types import *
 import tools.codegen.api.cpp as cpp
+from tools.codegen.api.cpp import CppSignature, CppSignatureGroup
 import tools.codegen.local as local
 from tools.codegen.gen import pythonify_default
 from tools.codegen.model import *
